// lib/api/generated/model/updateWorkingHourCommand.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */
import type { DayOfWeek } from "./dayOfWeek";

export interface UpdateWorkingHourCommand {
  dayOfWeek?: DayOfWeek;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  isBreak?: boolean;
  /** @nullable */
  staffId?: number | null;
}