// lib/api/generated/model/dayOfWeek.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * WAIC API
 * WAIC API endpoints
 * OpenAPI spec version: v1
 */

export type DayOfWeek = (typeof DayOfWeek)[keyof typeof DayOfWeek];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const DayOfWeek = {
  Sunday: "Sunday",
  Monday: "Monday",
  Tuesday: "Tuesday",
  Wednesday: "Wednesday",
  Thursday: "Thursday",
  Friday: "Friday",
  Saturday: "Saturday",
} as const;